adjusted statistics service to return raw data

+++ src/main/java/ch/uzh/csg/reimbursement/repository/ExpenseRepository.java
@@ -32,9 +32,9 @@ public interface ExpenseRepository extends JpaRepository<Expense, Integer> {
 	public Set<Expense> search(@Param("relevantUsers") List<User> relevantUsers, @Param("accountingText") String accountingText);
 
 	@Query("SELECT COUNT(e) FROM Expense e WHERE e.state = :state")
-	public double countByState(@Param("state") ExpenseState state);
+	public int countByState(@Param("state") ExpenseState state);
 
 	@Query("SELECT COUNT(e) FROM Expense e")
-	public double countExpenses();
+	public int countExpenses();
 
 }
